test cases for Dataframe.col.summarize
